requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
openai==1.51.0
huggingface_hub==0.25.2
httpx>=0.27,<0.28
reportlab==4.2.2
